Move `try_extract_request()` to util
